[v1.6.1 Update] 테스트용 데이터 추가 생성
1. 학년별로 유저 180명씩 생성
2. 후보는 컴퓨터학부, 철학과, 회계학과에서 한 팀씩 총 3팀
3. 투표는 학년별로 18명씩 총 72명 진행
4. 학과별로는 총 24명 진행(16명창선, 8명 반대)

+++ src/main/java/com/ggomak/vote/springboot/controller/VoteApiController.java
@@ -84,8 +84,8 @@ public class VoteApiController {
     }
 
     @GetMapping("/api/v3/gradevote")
-    public List<Integer> getGradeVoteResult(){
-        List<Integer> data = voteService.getGradeVoteResult();
+    public List<Long> getGradeVoteResult(){
+        List<Long> data = voteService.getGradeVoteResult();
 
         return data;
     }
